Move utils classes into other modules
